+++ froseai/datasets/__init__.py
@@ -1,0 +1,5 @@
+from .cassification import FedDatasetsClassification
+
+__all__ = [
+    "FedDatasetsClassification"
+]
